# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: TechniConnect - Electronics Technician & Shop Marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /auth/me/location:
    put:
      operationId: updateMyLocation
      tags: [auth]
      summary: Update current user location
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LocationInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # USERS
...
  /bookings:
    get:
      operationId: listBookings
      tags: [bookings]
      summary: List bookings
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: role
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Bookings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingListResponse"
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Book a technician
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{id}:
    get:
      operationId: getBooking
      tags: [bookings]
      summary: Get booking by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{id}/status:
    put:
      operationId: updateBookingStatus
      tags: [bookings]
      summary: Update booking status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingStatusInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{id}/location:
    put:
      operationId: updateBookingLocation
      tags: [bookings]
      summary: Update technician live location for active booking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LocationInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{id}/cancel:
    put:
      operationId: cancelBooking
      tags: [bookings]
      summary: Cancel booking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"

  # REVIEWS
...
  /admin/bookings:
    get:
      operationId: listAllBookings
      tags: [admin]
      summary: List all bookings (admin)
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Bookings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingListResponse"
...
components:
  schemas:
...
    LocationInput:
      type: object
      properties:
        lat:
          type: number
        lng:
          type: number
      required: [lat, lng]
...
    RegisterInput:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [user, technician, shop_owner]
        lat:
          type: ["number", "null"]
        lng:
          type: ["number", "null"]
        address:
          type: ["string", "null"]
      required: [name, email, password, role]
    LoginInput:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required: [token, user]
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        role:
          type: string
        lat:
          type: ["number", "null"]
        lng:
          type: ["number", "null"]
        address:
          type: ["string", "null"]
        profileImage:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
      required: [id, name, email, role, isActive, createdAt]
...
    Booking:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        technicianId:
          type: integer
        serviceId:
          type: ["integer", "null"]
        status:
          type: string
          enum: [pending, accepted, in_progress, completed, cancelled]
        scheduledAt:
          type: ["string", "null"]
        userLat:
          type: ["number", "null"]
        userLng:
          type: ["number", "null"]
        technicianLat:
          type: ["number", "null"]
        technicianLng:
          type: ["number", "null"]
        totalAmount:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        userName:
          type: ["string", "null"]
        technicianName:
          type: ["string", "null"]
        serviceName:
          type: ["string", "null"]
        userPhone:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, userId, technicianId, status, createdAt, updatedAt]
    BookingInput:
      type: object
      properties:
        technicianId:
          type: integer
        serviceId:
          type: integer
        scheduledAt:
          type: string
        userLat:
          type: number
        userLng:
          type: number
        totalAmount:
          type: number
        notes:
          type: string
      required: [technicianId]
    BookingStatusInput:
      type: object
      properties:
        status:
          type: string
          enum: [pending, accepted, in_progress, completed, cancelled]
      required: [status]
    BookingListResponse:
      type: object
      properties:
        bookings:
          type: array
          items:
            $ref: "#/components/schemas/Booking"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [bookings, total, page, limit]
...